llama : adapter
ggml-ci
