+++ common/common.h
@@ -481,7 +481,17 @@ std::string fs_get_cache_file(const std::string & filename);
 struct common_init_result {
     struct llama_model   * model   = nullptr;
     struct llama_context * context = nullptr;
+
     std::vector<common_lora_adapter_container> lora_adapters;
+
+    ~common_init_result() {
+        llama_free(context);
+        llama_free_model(model);
+
+        for (auto & lora_adapter : lora_adapters) {
+            llama_lora_adapter_free(lora_adapter.adapter);
+        }
+    }
 };
 
 struct common_init_result     common_init_from_params(common_params & params);
